+++ WukLamark/Models/MarkerShareEntry.cs
@@ -8,7 +8,7 @@ namespace WukLamark.Models
         public Guid? SourceId { get; set; }
         public string Name { get; set; } = "Unnamed Location";
         public Vector3 Position { get; set; }
-        public ushort TerritoryId { get; set; }
+        public uint TerritoryId { get; set; }
         public sbyte WardId { get; set; }
         public uint MapId { get; set; }
         public uint WorldId { get; set; }
