+++ WukLamark/Plugin.cs
@@ -210,7 +210,7 @@ public sealed class Plugin : IDalamudPlugin
     /// Called when the player changes territory (zone) in the game.
     /// -- Updates the current world ID in the LocationHelper to the current world ID.
     /// </summary>
-    private void OnTerritoryChange(ushort _) => LocationHelper.UpdateCurrentWorldId();
+    private void OnTerritoryChange(uint _) => LocationHelper.UpdateCurrentWorldId();
 
     /// <summary>
     /// Handles the /wlmark slash command with optional arguments.
